+++ CHANGES.rst
@@ -111,6 +111,7 @@
 - Substantial refactor of arsort.py
 - Substantial refactor of arsetup.py
 
+- Deprecates ['trace']['combine']['match'] > 0.0 option
 
 0.7 (2017-02-07)
 ----------------